pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-strands-agents-rag"
version = "0.1.0"
description = "AWS Strands Agents RAG with Ollama and Milvus"
requires-python = ">=3.10"
authors = [
    {name = "Claude Berube", email = "cberube@ipsofactointeractif.ca"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "numpy<2.0",
    "strands-agents>=1.27.0",
    "ollama>=0.1.0",
    "pymilvus>=2.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "tqdm>=4.65.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
